+++ rentacar/src/main/java/com/oguzhanturk/rentacar/dataAccess/abstracts/CarDao.java
@@ -1,5 +1,6 @@
 package com.oguzhanturk.rentacar.dataAccess.abstracts;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -9,7 +10,7 @@ import com.oguzhanturk.rentacar.entities.concretes.Car;
 
 @Repository
 public interface CarDao extends JpaRepository<Car, Integer> {
-	
-	List<Car> getByDailyPriceLessThanEqual(double dailyPrice);
+
+	List<Car> getByDailyPriceLessThanEqual(BigDecimal dailyPrice);
 
 }
